Combine type comment utilities into ``_type_comments.py``
